Move the copyDirectory function
